# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: StockPilot - Inventory & POS SaaS
  description: '

    ## StockPilot API - Inventory & POS SaaS


    Multi-tenant inventory management, point-of-sale, finance tracking, and AI insights

    for small businesses.


    ### Authentication

    All API endpoints require a Bearer token in the Authorization header:

    `Authorization: Bearer <access_token>`


    Include the business ID in the X-Business-Id header:

    `X-Business-Id: <business_id>`


    ### Rate Limiting

    - Default: 100 requests per minute per user

    - Auth endpoints: 5 requests per minute per IP


    ### Support

    For issues and feature requests, contact support.

    '
  version: 1.0.0
paths:
...
  /api/v1/purchases:
    get:
      tags:
      - purchases
      summary: List All
      operationId: list_all_api_v1_purchases_get
      security:
      - HTTPBearer: []
      parameters:
      - name: supplier_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Supplier Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PurchaseOut'
                title: Response List All Api V1 Purchases Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - purchases
      summary: Create
      description: 'FR-10: multi-item purchase. Totals auto-calculated, stock up on confirm.'
      operationId: create_api_v1_purchases_post
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/sales:
    get:
      tags:
      - sales
      - sales
      summary: List Sales
      description: 'FR-12.8: search/filter sales (batched, default 50 — not 200).'
      operationId: list_sales_api_v1_sales_get
      security:
      - HTTPBearer: []
      parameters:
      - name: customer_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Customer Id
      - name: payment_status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Payment Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SaleOut'
                title: Response List Sales Api V1 Sales Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/expenses:
    get:
      tags:
      - expenses
      summary: List All
      description: 'FR-15.5: expense history with optional category/date filters.'
      operationId: list_all_api_v1_expenses_get
      security:
      - HTTPBearer: []
      parameters:
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
      - name: date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date From
      - name: date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date To
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExpenseOut'
                title: Response List All Api V1 Expenses Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - expenses
      summary: Create
      description: 'FR-15.1/15.2/15.3: create expense record.'
      operationId: create_api_v1_expenses_post
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/reports/sales:
    get:
      tags:
      - reports
      summary: Sales Report
      description: 'FR-23.1/23.6/23.7/23.8 (+Excel/PDF): sales report.'
      operationId: sales_report_api_v1_reports_sales_get
      security:
      - HTTPBearer: []
      parameters:
      - name: preset
        in: query
        required: false
        schema:
          type: string
          default: month
          title: Preset
      - name: date_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date From
      - name: date_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Date To
      - name: format
        in: query
        required: false
        schema:
          type: string
          default: json
          title: Format
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100000
          minimum: 1
          default: 2000
          title: Limit
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reports/inventory:
    get:
      tags:
      - reports
      summary: Inventory Report
      description: 'FR-23.2/23.7/23.8: inventory report + CSV/Excel/PDF export.'
      operationId: inventory_report_api_v1_reports_inventory_get
      security:
      - HTTPBearer: []
      parameters:
      - name: format
        in: query
        required: false
        schema:
          type: string
          default: json
          title: Format
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100000
          minimum: 1
          default: 5000
          title: Limit
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reports/purchases:
    get:
      tags:
      - reports
      summary: Purchase Report
      description: 'FR-23.3/23.7/23.8: purchase report.'
      operationId: purchase_report_api_v1_reports_purchases_get
      security:
      - HTTPBearer: []
      parameters:
      - name: preset
        in: query
        required: false
        schema:
          type: string
          default: month
          title: Preset
      - name: format
        in: query
        required: false
        schema:
          type: string
          default: json
          title: Format
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100000
          minimum: 1
          default: 2000
          title: Limit
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reports/expenses:
    get:
      tags:
      - reports
      summary: Expense Report
      description: 'FR-23.4/23.7/23.8: expense report.'
      operationId: expense_report_api_v1_reports_expenses_get
      security:
      - HTTPBearer: []
      parameters:
      - name: preset
        in: query
        required: false
        schema:
          type: string
          default: month
          title: Preset
      - name: format
        in: query
        required: false
        schema:
          type: string
          default: json
          title: Format
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100000
          minimum: 1
          default: 2000
          title: Limit
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reports/profit:
    get:
      tags:
      - reports
      summary: Profit Report
      description: 'FR-23.5/23.7/23.8: profit report.'
      operationId: profit_report_api_v1_reports_profit_get
      security:
      - HTTPBearer: []
      parameters:
      - name: preset
        in: query
        required: false
        schema:
          type: string
          default: month
          title: Preset
      - name: format
        in: query
        required: false
        schema:
          type: string
          default: json
          title: Format
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reports/profit/pdf:
    get:
      tags:
      - reports
      summary: Profit Pdf
      description: 'FR-23.8: profit report as PDF download.'
      operationId: profit_pdf_api_v1_reports_profit_pdf_get
      security:
      - HTTPBearer: []
      parameters:
      - name: preset
        in: query
        required: false
        schema:
          type: string
          default: month
          title: Preset
      - name: X-Business-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Business-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ExpenseCreate:
      properties:
        category:
          type: string
          maxLength: 50
          minLength: 1
          title: Category
        amount:
          type: number
          exclusiveMinimum: 0.0
          title: Amount
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        expense_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expense Date
        payment_method:
          type: string
          title: Payment Method
          default: cash
      type: object
      required:
      - category
      - amount
      title: ExpenseCreate
    ExpenseOut:
      properties:
        id:
          type: integer
          title: Id
        category:
          type: string
          title: Category
        amount:
          type: number
          title: Amount
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        expense_date:
          type: string
          format: date-time
          title: Expense Date
        payment_method:
          type: string
          title: Payment Method
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - category
      - amount
      - expense_date
      - payment_method
      - created_at
      title: ExpenseOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PurchaseCreate:
      properties:
        supplier_id:
          type: integer
          title: Supplier Id
        items:
          items:
            $ref: '#/components/schemas/PurchaseItemIn'
          type: array
          minItems: 1
          title: Items
        discount_amount:
          type: number
          title: Discount Amount
          default: 0
        tax_amount:
          type: number
          title: Tax Amount
          default: 0
        paid_amount:
          type: number
          title: Paid Amount
          default: 0
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - supplier_id
      - items
      title: PurchaseCreate
    PurchaseItemIn:
      properties:
        product_id:
          type: integer
          title: Product Id
        quantity:
          type: integer
          exclusiveMinimum: 0.0
          title: Quantity
        unit_cost:
          type: number
          minimum: 0.0
          title: Unit Cost
      type: object
      required:
      - product_id
      - quantity
      - unit_cost
      title: PurchaseItemIn
    PurchaseItemOut:
      properties:
        id:
          type: integer
          title: Id
        product_id:
          type: integer
          title: Product Id
        product_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Name
        quantity:
          type: integer
          title: Quantity
        unit_cost:
          type: number
          title: Unit Cost
        line_total:
          type: number
          title: Line Total
      type: object
      required:
      - id
      - product_id
      - quantity
      - unit_cost
      - line_total
      title: PurchaseItemOut
    PurchaseOut:
      properties:
        id:
          type: integer
          title: Id
        supplier_id:
          type: integer
          title: Supplier Id
        supplier_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Supplier Name
        purchase_date:
          type: string
          format: date-time
          title: Purchase Date
        subtotal:
          type: number
          title: Subtotal
        discount_amount:
          type: number
          title: Discount Amount
        tax_amount:
          type: number
          title: Tax Amount
        total_amount:
          type: number
          title: Total Amount
        paid_amount:
          type: number
          title: Paid Amount
        payment_status:
          type: string
          title: Payment Status
        status:
          type: string
          title: Status
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        items:
          items:
            $ref: '#/components/schemas/PurchaseItemOut'
          type: array
          title: Items
          default: []
      type: object
      required:
      - id
      - supplier_id
      - purchase_date
      - subtotal
      - discount_amount
      - tax_amount
      - total_amount
      - paid_amount
      - payment_status
      - status
      title: PurchaseOut
...
    SaleItemOut:
      properties:
        id:
          type: integer
          title: Id
        product_id:
          type: integer
          title: Product Id
        product_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Product Name
        quantity:
          type: integer
          title: Quantity
        unit_price:
          type: number
          title: Unit Price
        discount:
          type: number
          title: Discount
        line_total:
          type: number
          title: Line Total
        returned_qty:
          type: integer
          title: Returned Qty
      type: object
      required:
      - id
      - product_id
      - quantity
      - unit_price
      - discount
      - line_total
      - returned_qty
      title: SaleItemOut
    SaleOut:
      properties:
        id:
          type: integer
          title: Id
        invoice_no:
          type: string
          title: Invoice No
        customer_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Customer Id
        customer_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Customer Name
        subtotal:
          type: number
          title: Subtotal
        discount_amount:
          type: number
          title: Discount Amount
        tax_percent:
          type: number
          title: Tax Percent
        tax_amount:
          type: number
          title: Tax Amount
        total_amount:
          type: number
          title: Total Amount
        paid_amount:
          type: number
          title: Paid Amount
        payment_method:
          type: string
          title: Payment Method
        payment_status:
          type: string
          title: Payment Status
        status:
          type: string
          title: Status
        created_at:
          type: string
          format: date-time
          title: Created At
        items:
          items:
            $ref: '#/components/schemas/SaleItemOut'
          type: array
          title: Items
          default: []
      type: object
      required:
      - id
      - invoice_no
      - subtotal
      - discount_amount
      - tax_percent
      - tax_amount
      - total_amount
      - paid_amount
      - payment_method
      - payment_status
      - status
      - created_at
      title: SaleOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer